Implemented feature request for Bug #1641. Munge Process priority is now specified by the order of munge processes in the munge process folder. This is done from the munge process group editor now, by selecting the munge process in the table, and using the arrow buttons at the right side to move them up and down the list. The match engine will process the munge processes in this order. As such, the JSpinner in the munge process editor is gone. A few other changes were made:
* The match priority is now kept as an Integer, instead of a Short, since the munge process' index in the munge process folder is returned as an int. Fortunately, the match percent column in the repository schema should be large enough so no schema changes should be necessary.

* The Hibernate configuration for Project returns Munge Processes sorted by the match priority, so when munge processes are loaded with a project, they are placed in proper order in the folder.

* MatchMakerFolder passes an isCompound flag when firing an addChild event to prevent one of the listeners in MatchMakerSwingSession (SelectRemovedEditorListener? or something like that), from selecting the moved munge process editor when simply moving munge processes around, since this makes change the order of the munge processes really annoying.

+++ trunk/src/ca/sqlpower/matchmaker/dao/xml/ProjectSAXHandler.java
@@ -386,7 +386,7 @@ public class ProjectSAXHandler extends DefaultHandler {
                     } else if (aname.equals("colour")) {
                         process.setColour(Color.decode(aval));
                     } else if (aname.equals("priority")) {
-                        process.setMatchPriority(Short.valueOf(aval));
+                        process.setMatchPriority(Integer.valueOf(aval));
                     } else {
                         logger.warn("Unexpected attribute of <munge-process>: " + aname + "=" + aval + " at " + locator);
                     }
